requirements.txt
streamlit>=1.36.0
pypdf>=4.2.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.4
scikit-learn>=1.5.1
rank-bm25>=0.2.2
openai>=1.40.0
transformers>=4.43.3
torch>=2.3.1
python-dotenv>=1.0.1
fastembed>=0.3.7
